project4-multiagent/requirements.txt
# Core dependencies
pyautogen==0.10.0
openai>=1.0.0
python-dotenv>=1.0.0

# Web interface
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
rich>=13.0.0
